rfvp::subsystem::components::syscalls

Module graph

Structs

GaijiLoad

Arg1: code (string) - The character or symbol that should be substituted by a gaiji glyph. - Example: “怒”, “汗”, “♪” Arg2: size slot (integer 12–64) - The text-size slot where this gaiji mapping applies. Arg3: path (string) - NVSG gaiji glyph resource path. - Example: “graph/gaiji_ikari”

GraphLoad

Load or unload a graphical texture into the engine.

GraphRGB

Adjust the color tone of a loaded texture.

PrimExitGroup

Set the root primitive index The engine will begin rendering from this primitive. Arg1: the primitive, which should not larger that 0x1000

PrimGroupIn

Add a primitive as a child of another primitive (group insertion).

PrimGroupMove

Move a primitive in the scene hierarchy relative to another primitive.

PrimGroupOut

Remove a primitive from the scene hierarchy.

PrimHit

Check if a sprite primitive was clicked or “hit”.

PrimSetAlpha

Set the alpha (transparency) value of the corresponding primitive

PrimSetBlend

Set the blending mode flag of the corresponding primitive

PrimSetDraw

Set the draw flag of the corresponding primitive Arg1: primitive index Arg2: draw flag

PrimSetNull

Reset (clear) the corresponding primitive

PrimSetOP

Set the operation (position) of the corresponding primitive

PrimSetRS

Set the rotation and scale factor of the corresponding primitive

PrimSetRS2

Set the rotation and independent scale factors of the corresponding primitive

PrimSetSnow

PrimSetSprt

Initialize a primitive as a sprite and bind it to a texture or special source.

PrimSetText

Initialize a primitive as a text placeholder bound to a text object.

PrimSetTile

Initialize a primitive as a solid-color tile

PrimSetUV

Set the UV sub-rectangle of a primitive

PrimSetWH

Set the width and height of a primitive’s texture region

PrimSetXY

PrimSetZ

Set the Z-depth of a sprite primitive.

Functions

gaiji_load

graph_load

graph_rgb

prim_exit_group

prim_group_in

prim_group_move

prim_group_out

prim_hit

prim_set_alpha

set the primitive’s alpha value

prim_set_blend

prim_set_draw

prim_set_null

reset the primitive

prim_set_op

prim_set_rs

set the primitive’s rotation and scale, and the scale value is the same in x and y

prim_set_rs2

set the primitive’s rotation and scale, and the scale value is different in x and y

prim_set_snow

prim_set_sprt

prim_set_text

prim_set_tile

prim_set_uv

prim_set_wh

prim_set_xy

prim_set_z