rfvp::subsystem::components::syscalls

Module input

Structs

ControlMask

Enable/disable control and shift keys, which are used to speed up text display in AVG games. When control is masked, both control and shift keys are disabled. Arg1: mask (nil to enable, non-nil to disable)

ControlPulse

Skip mode in AVG games No arguments

InputFlash

Refresh all input states, such as which keys are pressed No arguments

InputGetCursIn

Get whether the cursor is within the window No arguments Returns true if the cursor is inside the window, nil otherwise

InputGetCursX

Get the X coordinate of the cursor, relative to the window No arguments Returns an integer value representing the X coordinate

InputGetCursY

Get the Y coordinate of the cursor, relative to the window No arguments Returns an integer value representing the Y coordinate

InputGetDown

FVP’s keycode: #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)] pub enum KeyCode { Shift = 0, Ctrl = 1, LeftClick = 2, // virtual RightClick = 3, // virtual MouseLeft = 4, MouseRight = 5, Esc = 6, Enter = 7, Space = 8, UpArrow = 9, DownArrow = 10, LeftArrow = 11, RightArrow = 12, F1 = 13, F2 = 14, F3 = 15, F4 = 16, F5 = 17, F6 = 18, F7 = 19, F8 = 20, F9 = 21, F10 = 22, F11 = 23, F12 = 24, Tab = 25, }

InputGetEvent

Get the next input event from the queue No arguments Returns a table with the following structure if an event is available: { keycode, x, y } where keycode is an integer representing the keycode, and x, y are the cursor coordinates at the time of the event. Returns nil if no event is available.

InputGetRepeat

Get the keys that are currently being repeated Notice that both wheel movements and button repeats will be reset to 0 during each frame update. No arguments Returns an integer value representing the keycode, or 0 if no key is being repeated

InputGetState

Get the current state of all keys, such as which keys are pressed No arguments Returns an integer value representing the bitmask of the current key states

InputGetUp

Get the keycode of the most recently released key No arguments Returns an integer value representing the keycode, or nil if no key has been released

InputGetWheel

Get mouse wheel value (usually for scrolling) Notice that the wheel value will be reset to 0 during each frame update. No arguments Returns an integer value representing the wheel movement

InputSetClick

Set the click state from script side. Maybe this is used to simulate mouse clicks in certain scenarios. Arg1: clicked (0 for not clicked, 1 for clicked)

Functions

control_mask

control_pulse

Skip mode in AVG games

input_flash

input_get_curs_in

input_get_curs_x

input_get_curs_y

input_get_down

input_get_event

input_get_repeat

input_get_state

input_get_up

input_get_wheel

input_set_click