rfvp::subsystem::components::syscalls

Module sound

Structs

AudioLoad

AudioPlay

AudioSilentOn

AudioSlientOn

AudioState

AudioStop

AudioType

AudioVol

SoundLoad

SoundMasterVol

SoundPlay

SoundSilentOn

SoundSlientOn

SoundStop

SoundType

SoundTypeVol

SoundVol

SoundVolume

Functions

audio_load

load audio on a specific channel, used for voice and sound effects

audio_play

play audio on a specific channel, and loop it if needed

audio_silent_on

Silence audio on a specific channel (BGM).

audio_slient_on

Backward-compatible misspelling.

audio_state

test the specific channel is playing or not

audio_stop

stop audio on a specific channel

audio_type

audio_vol

sound_load

load sound on a specific channel, used for voice and sound effects

sound_master_vol

sound_play

sound_silent_on

Silence sound on a specific channel (SE).

sound_slient_on

Backward-compatible misspelling.

sound_stop

sound_type

sound_type_vol

sound_volume