rfvp::subsystem::components::syscalls

Module text

Structs

TextBuff

TextClear

TextColor

TextFont

TextFontCount

TextFontGet

TextFontName

TextFontSet

TextFormat

TextFunction

TextOutSize

TextPause

TextPos

TextPrint

TextRepaint

TextRepaint() IDA SYSCALL_SPECS: argc=0

TextReprint

TextShadowDist

TextSize

TextSkip

TextSpace

TextSpeed

TextSuspendChr

TextTest

Functions

text_buff

text_clear

text_color

text_font

text_font_count

text_font_get

text_font_name

text_font_set

text_format

text_function

text_out_size

text_pause

text_pos

text_print

text_reprint

text_shadow_dist

text_size

text_skip

text_space

text_speed

set the text speed for the corresponding text id 0 is for immediate display

text_suspend_chr

set the kinsoku chars (禁则) for the corresponding text id

text_test

test the const string was readed the original implementation use a bitmap to record the text address