rfvp::subsystem::resources::input_manager

Enum KeyCode

pub enum KeyCode {

    Shift = 0,
    Ctrl = 1,
    LeftClick = 2,
    RightClick = 3,
    MouseLeft = 4,
    MouseRight = 5,
    Esc = 6,
    Enter = 7,
    Space = 8,
    UpArrow = 9,
    DownArrow = 10,
    LeftArrow = 11,
    RightArrow = 12,
    F1 = 13,
    F2 = 14,
    F3 = 15,
    F4 = 16,
    F5 = 17,
    F6 = 18,
    F7 = 19,
    F8 = 20,
    F9 = 21,
    F10 = 22,
    F11 = 23,
    F12 = 24,
    Tab = 25,

}

Key codes used by FVP, auctually keycode is just the index of the bit in input_state input_state |= 1 << keycode if input_state is zero, then no key is pressed This term is little confusing, but I keep it for compatibility

Variants

Shift = 0

Ctrl = 1

LeftClick = 2

RightClick = 3

MouseLeft = 4

MouseRight = 5

Esc = 6

Enter = 7

Space = 8

UpArrow = 9

DownArrow = 10

LeftArrow = 11

RightArrow = 12

F1 = 13

F2 = 14

F3 = 15

F4 = 16

F5 = 17

F6 = 18

F7 = 19

F8 = 20

F9 = 21

F10 = 22

F11 = 23

F12 = 24

Tab = 25