rfvp::subsystem::scene

Trait Scene

pub trait Scene: Send + Sync {
    // Provided methods
    fn on_start(&mut self, _data: &mut GameData) { ... }
    fn on_update(&mut self, _data: &mut GameData) { ... }
    fn late_update(&mut self, _data: &mut GameData) { ... }
    fn on_stop(&mut self, _data: &mut GameData) { ... }
}

Trait to implement in order to define a Scene.

Note: The engine runs the VM on a dedicated OS thread and shares global state via Arc<RwLock<Box<GameData>>>. GameData contains a SceneController which may temporarily hold a pending scene transition (Box<dyn Scene>). To keep GameData Send + Sync (so it can be wrapped by RwLock and shared across threads), all scene implementations must be thread-safe.

Provided Methods

fn on_start(&mut self, _data: &mut GameData)

Will be called once before the new game loop iteration. Useful to initialize resources and add everything you need in the world.

fn on_update(&mut self, _data: &mut GameData)

Will be called each game loop, before the systems execution

fn late_update(&mut self, _data: &mut GameData)

Will be called each game loop, after the systems execution

fn on_stop(&mut self, _data: &mut GameData)

Will be called for deleted scene at the end of the frame where it was deleted