rfvp::utils::time

Struct Time

pub struct Time { /* private fields */ }

A clock that tracks how much it has advanced (and how much real time has elapsed) since its previous update and since its creation.

Implementations

impl Time

pub fn new(startup: Instant) -> Self

Constructs a new Time instance with a specific startup Instant.

pub fn update(&mut self)

Updates the internal time measurements.

pub fn update_with_instant(&mut self, instant: Instant)

Updates time with a specified Instant.

This method is provided for use in tests. Calling this method as part of your app will most likely result in inaccurate timekeeping.

pub fn startup(&self) -> Instant

Returns the Instant the clock was created.

This usually represents when the app was started.

pub fn first_update(&self) -> Option<Instant>

Returns the Instant when update was first called, if it exists.

This usually represents when the first app update started.

pub fn last_update(&self) -> Option<Instant>

Returns the Instant when update was last called, if it exists.

This usually represents when the current app update started.

pub fn delta(&self) -> Duration

Returns how much time has advanced since the last update, as a Duration.

pub fn delta_seconds(&self) -> f32

Returns how much time has advanced since the last update, as f32 seconds.

pub fn delta_seconds_f64(&self) -> f64

Returns how much time has advanced since the last update, as f64 seconds.

pub fn elapsed(&self) -> Duration

Returns how much time has advanced since startup, as Duration.

pub fn elapsed_seconds(&self) -> f32

Returns how much time has advanced since startup, as f32 seconds.

Note: This is a monotonically increasing value. It’s precision will degrade over time. If you need an f32 but that precision loss is unacceptable, use elapsed_seconds_wrapped.

pub fn elapsed_seconds_f64(&self) -> f64

Returns how much time has advanced since startup, as f64 seconds.

pub fn elapsed_wrapped(&self) -> Duration

Returns how much time has advanced since startup modulo the wrap_period, as Duration.

pub fn elapsed_seconds_wrapped(&self) -> f32

Returns how much time has advanced since startup modulo the wrap_period, as f32 seconds.

This method is intended for applications (e.g. shaders) that require an f32 value but suffer from the gradual precision loss of elapsed_seconds.

pub fn elapsed_seconds_wrapped_f64(&self) -> f64

Returns how much time has advanced since startup modulo the wrap_period, as f64 seconds.

pub fn raw_delta(&self) -> Duration

Returns how much real time has elapsed since the last update, as a Duration.

pub fn raw_delta_seconds(&self) -> f32

Returns how much real time has elapsed since the last update, as f32 seconds.

pub fn raw_delta_seconds_f64(&self) -> f64

Returns how much real time has elapsed since the last update, as f64 seconds.

pub fn raw_elapsed(&self) -> Duration

Returns how much real time has elapsed since startup, as Duration.

pub fn raw_elapsed_seconds(&self) -> f32

Returns how much real time has elapsed since startup, as f32 seconds.

Note: This is a monotonically increasing value. It’s precision will degrade over time. If you need an f32 but that precision loss is unacceptable, use raw_elapsed_seconds_wrapped.

pub fn raw_elapsed_seconds_f64(&self) -> f64

Returns how much real time has elapsed since startup, as f64 seconds.

pub fn raw_elapsed_wrapped(&self) -> Duration

Returns how much real time has elapsed since startup modulo the wrap_period, as Duration.

pub fn raw_elapsed_seconds_wrapped(&self) -> f32

Returns how much real time has elapsed since startup modulo the wrap_period, as f32 seconds.

This method is intended for applications (e.g. shaders) that require an f32 value but suffer from the gradual precision loss of raw_elapsed_seconds.

pub fn raw_elapsed_seconds_wrapped_f64(&self) -> f64

Returns how much real time has elapsed since startup modulo the wrap_period, as f64 seconds.

pub fn wrap_period(&self) -> Duration

Returns the modulus used to calculate elapsed_wrapped and raw_elapsed_wrapped.

Note: The default modulus is one hour.

pub fn set_wrap_period(&mut self, wrap_period: Duration)

Sets the modulus used to calculate elapsed_wrapped and raw_elapsed_wrapped.

Note: This will not take effect until the next update.

Panics

Panics if wrap_period is a zero-length duration.

pub fn relative_speed(&self) -> f32

Returns the speed the clock advances relative to your system clock, as f32. This is known as “time scaling” or “time dilation” in other engines.

Note: This function will return zero when time is paused.

pub fn relative_speed_f64(&self) -> f64

Returns the speed the clock advances relative to your system clock, as f64. This is known as “time scaling” or “time dilation” in other engines.

Note: This function will return zero when time is paused.

pub fn set_relative_speed(&mut self, ratio: f32)

Sets the speed the clock advances relative to your system clock, given as an f32.

For example, setting this to 2.0 will make the clock advance twice as fast as your system clock.

Note: This does not affect the raw_* measurements.

Panics

Panics if ratio is negative or not finite.

pub fn set_relative_speed_f64(&mut self, ratio: f64)

Sets the speed the clock advances relative to your system clock, given as an f64.

For example, setting this to 2.0 will make the clock advance twice as fast as your system clock.

Note: This does not affect the raw_* measurements.

Panics

Panics if ratio is negative or not finite.

pub fn pause(&mut self)

Stops the clock, preventing it from advancing until resumed.

Note: This does affect the raw_* measurements.

pub fn unpause(&mut self)

Resumes the clock if paused.

pub fn is_paused(&self) -> bool

Returns true if the clock is currently paused.